fix(eval): restore CSV export architecture boundary
