+++ src/csv.ts
@@ -310,9 +310,9 @@ export function testCaseFromCsvRow(row: CsvRow): TestCase {
 export function serializeObjectArrayAsCSV(vars: object[]): string {
   invariant(vars.length > 0, 'No variables to serialize');
   const columnNames = Object.keys(vars[0]).join(',');
-  // NOTE: values here are intentionally NOT formula-escaped (see escapeCsvFormula).
-  // This output is a generated test-case dataset meant to be re-imported via the
-  // CSV test loader, and prefixing a cell would corrupt vars on round-trip.
+  // This generated test-case dataset is intentionally not formula-escaped because
+  // prefixing a cell would corrupt vars on round-trip. Formula escaping is applied
+  // only by the eval result export path.
   const rows = vars
     .map(
       (result) =>
@@ -322,51 +322,4 @@ export function serializeObjectArrayAsCSV(vars: object[]): string {
     )
     .join('\n');
   return [columnNames, rows].join('\n') + '\n';
-}
-
-function isNumericCell(value: string): boolean {
-  // Only finite numbers are safe to leave unescaped — `Number('-Infinity')` is a
-  // valid (non-NaN) number, so guard with isFinite so "-Infinity"/"+Infinity"
-  // (and overflow like "-1e400") are still treated as formula triggers.
-  return value.trim() !== '' && Number.isFinite(Number(value));
-}
-
-/**
- * Neutralize spreadsheet formula (CSV) injection — CWE-1236, OWASP "CSV Injection".
- *
- * Excel, Google Sheets and LibreOffice execute a cell as a formula when its text
- * begins with `=`, `+`, `-`, `@`, a TAB or a carriage return. promptfoo's exported
- * eval/redteam results contain model output, test vars and grader comments —
- * attacker-influenced, especially under red teaming — so an exported CSV opened in
- * a spreadsheet could run code, exfiltrate data, or phish via a hyperlink.
- * csv-stringify's quoting does NOT prevent this; the cell value itself must be
- * prefixed.
- *
- * We prepend a single apostrophe, the spreadsheet-native "treat the rest as text"
- * marker (it is not shown in the cell). To stay minimally disruptive we only touch
- * genuinely dangerous values: a leading `-`/`+` on an otherwise-numeric value
- * (e.g. "-5", "+3.14", a score) is left alone because spreadsheets read those as
- * numbers, not formulas. Empty strings are returned unchanged.
- *
- * Apply this only at export time. promptfoo's CSV import path and any user-authored
- * CSVs must never be mutated, so round-trips stay lossless.
- */
-export function escapeCsvFormula(value: string): string {
-  if (value.length === 0) {
-    return value;
-  }
-  // Spreadsheets strip leading whitespace AND zero-width characters before parsing,
-  // so inspect the first visible character after removing both (this also covers
-  // leading TAB / CR obfuscation). JS `\s` already covers BOM/NBSP but misses the
-  // zero-width space/non-joiner/joiner (U+200B–U+200D), so strip those explicitly.
-  const unpadded = value.replace(/^[\s\u200B\u200C\u200D]+/, '');
-  const firstChar = unpadded[0];
-  if (firstChar === undefined) {
-    return value;
-  }
-  const isFormulaTrigger =
-    firstChar === '=' ||
-    firstChar === '@' ||
-    ((firstChar === '-' || firstChar === '+') && !isNumericCell(unpadded));
-  return isFormulaTrigger ? `'${value}` : value;
 }
